+++ src/lib/fileio.c
@@ -63,7 +63,7 @@ static int items;
 
 #ifndef DOUBLE
 
-int fwritef (double *ptr, const unsigned int size, const int nitems, FILE *fp)
+int fwritef (double *ptr, const size_t size, const int nitems, FILE *fp)
 {
    int i;
    if (items < nitems) {
@@ -78,7 +78,7 @@ int fwritef (double *ptr, const unsigned int size, const int nitems, FILE *fp)
    return fwrite(f, sizeof(float), nitems, fp);
 }
 
-int freadf (double *ptr, const unsigned int size, const int nitems, FILE *fp)
+int freadf (double *ptr, const size_t size, const int nitems, FILE *fp)
 {
    int i, n;
    if (items < nitems) {
